requirements.txt
argparse==1.2.1
python-keystoneclient==1.2.0
python-novaclient==2.22.0
